Bug fixes in OpenCL backend

Also, includes af_print implementation for OpenCL backend.

+++ include/af/defines.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stdlib.h>
+
 typedef enum {
     AF_SUCCESS=0,
     AF_ERR_INTERNAL,
@@ -52,7 +54,7 @@ typedef enum {
   #include <stdbool.h>
 #endif
 
-typedef long long dim_type;
+typedef size_t dim_type;
 
 #include <cstddef>
 typedef struct af_seq {
